utils/pyproject.toml
[project]
name = "utils"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "linhsolar", email = "linh.truong@aalto.fi" }
]
dependencies = [
    "paho-mqtt>=2.1.0",
    "polars>=1.22.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []
